+++ lib/rake_announcer.rb
@@ -7,11 +7,16 @@ require_relative "rake_announcer/version"
 module RakeAnnouncer
   class Error < StandardError; end
 
-  def self.announce_rake_task(task_name)
+  def self.announce_rake_task(task_name, prepend: false)
     announce_task = Rake::Task.define_task("announce_#{task_name}") do
       RakeAnnouncer.announce("Running #{task_name}")
     end
-    Rake::Task[task_name].enhance([announce_task.name])
+
+    if prepend
+      Rake::Task[task_name].prereqs.prepend(announce_task.name)
+    else
+      Rake::Task[task_name].enhance([announce_task.name])
+    end
   end
 
   def self.announce(message)