+++ lib/rake_announcer/rake_task.rb
@@ -5,12 +5,12 @@ require "rake_announcer"
 
 module RakeAnnouncer
   class RakeTask < Rake::TaskLib
-    def initialize(name: :rake_announcer, tasks: [])
+    def initialize(name: :rake_announcer, tasks: [], prepend: false)
       super()
 
       namespace name do
         tasks.each do |t|
-          RakeAnnouncer.announce_rake_task(t) unless t.to_s.start_with?(name.to_s)
+          RakeAnnouncer.announce_rake_task(t, prepend:) unless t.to_s.start_with?(name.to_s)
         end
 
         desc "Print the 'ALL TESTS PASSED' message"