+++ spec/rake_announcer_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe RakeAnnouncer do
 
   describe ".announce_rake_task" do
     before do
-      Rake::Task.define_task("dummy_task")
+      Rake::Task.define_task("dummy_task").clear
     end
 
     it "defines a new rake task" do
@@ -23,6 +23,22 @@ RSpec.describe RakeAnnouncer do
     it "prints an announcement when the original task is invoked" do
       described_class.announce_rake_task("dummy_task")
       expect { Rake::Task["dummy_task"].invoke }.to output(/Running dummy_task/).to_stdout
+    end
+
+    it "appends announce tasks by default" do
+      Rake::Task["dummy_task"].enhance(["pre_task"])
+      expect { described_class.announce_rake_task("dummy_task") }
+        .to change { Rake::Task["dummy_task"].prereqs }
+        .from(%w[pre_task])
+        .to(%w[pre_task announce_dummy_task])
+    end
+
+    it "can optionally prepend announce tasks" do
+      Rake::Task["dummy_task"].enhance(["pre_task"])
+      expect { described_class.announce_rake_task("dummy_task", prepend: true) }
+        .to change { Rake::Task["dummy_task"].prereqs }
+        .from(%w[pre_task])
+        .to(%w[announce_dummy_task pre_task])
     end
   end