+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* feat: Add option to prepend announce tasks
 * fix: Prevent own tasks from being announced
 * fix: Rename RakeAnnouncer.enhance_rake_task to RakeAnnouncer.announce_rake_task